requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.3
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
pytest==8.3.3
httpx==0.28.1
rapidfuzz==3.9.0
python-jose[cryptography]
python-multipart
pytest 
httpx 
pytest-asyncio